admin-portal/api/requirements.txt
fastapi
uvicorn
google-cloud-firestore
pydantic
python-multipart
httpx
